Teach the older fakes the two questions recognition asks
The Phase 4 cloud-index fake had never heard of a content hash, and one test
helper passed a message id as an Int. Both are the compile telling the truth
about an interface that grew, so the fake answers the new members rather than
the interface shrinking back.

Also says what `markBackedUp` now accepts as evidence, and that Phase 6's hash
matching is what catches a send interrupted after Telegram committed it.

+++ app/src/test/java/com/lumovault/app/data/remote/telegram/TdCloudMappingTest.kt
@@ -249,7 +249,7 @@ class TdCloudMappingTest {
 
     @Test
     fun aMalformedOrFutureManifestIsReadAsNothingRatherThanAsAPartialHash() {
-        fun withCaption(id: Int, text: String) = TdCloudMapper.toCloudMedia(
+        fun withCaption(id: Long, text: String) = TdCloudMapper.toCloudMedia(
             message(
                 id = id,
                 content = TdApi.MessagePhoto().apply {
